+++ domain/user/src/main/kotlin/io/github/siyual_park/user/domain/UserRemover.kt
@@ -5,6 +5,7 @@ import io.github.siyual_park.user.entity.User
 import io.github.siyual_park.user.repository.UserCredentialRepository
 import io.github.siyual_park.user.repository.UserRepository
 import io.github.siyual_park.user.repository.UserScopeRepository
+import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.stereotype.Component
 import org.springframework.transaction.reactive.TransactionalOperator
 import org.springframework.transaction.reactive.executeAndAwait
@@ -22,6 +23,11 @@ class UserRemover(
     }
 
     suspend fun remove(userId: Long, soft: Boolean = true) = operator.executeAndAwait {
+        val user = userRepository.findById(userId)
+        if (user == null || user.deletedAt != null) {
+            throw EmptyResultDataAccessException(1)
+        }
+
         userScopeRepository.deleteAllByUserId(userId)
         userCredentialRepository.deleteByUserId(userId)
         if (!soft) {
